+++ src/components/burger-constructor/burger-constructor.module.css
@@ -49,14 +49,3 @@
   align-items: center;
   gap: 5px;
 }
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.title {
-  text-shadow: 0px 0px 16px rgba(51, 51, 255, 0.25), 0px 0px 8px rgba(51, 51, 255, 0.25),
-    0px 4px 32px rgba(51, 51, 255, 0.5);
-}
