+++ src/components/burger-ingredient/burger-ingredient.module.css
@@ -11,21 +11,3 @@
 .image {
   cursor: pointer;
 }
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-items: center;
-  width: 100%;
-}
-
-.structure {
-  display: flex;
-  gap: 30px;
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  text-align: center;
-}
